major cleanup. moved simulatedArduino class to main. updated error handling of arduino connection
